Add the ability to allow a cache to be refreshed before it expires
In the case where lots of requests are coming in for an expensive endpoint it
is useful to begin the process of refreshing the cache before the expiry.
Otherwise you end can end up performing a number of expensive calls if multiple
calls come in after the cache expires but before the first one can return.